+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - **airbrake:** Use `ForEachCause` to find an HTTP request embedded into an error
 
+### Removed
+
+- `Causer` interface
+
 
 ## [0.7.1] - 2018-04-27